feat authorization activity:\ - update post_service.js, post_controller, post_routes, inedx.js\ - create a errorHandler_middleware.js

+++ src/services/post_service.js
@@ -63,24 +63,33 @@ export const createPost = async (postData, authorId) => {
     }
 };
 
-export const updatePost = async (id, postData) => {
+export const updatePost = async (id, postData, userId) => {
     const {title, content} = postData;
-    const [result] = await pool.query(
+
+    const post = await getPostById(id);
+
+    if (post.authorId !== userId) {
+        throw new ApiError(403, "Forbidden: You do not have permission to edit this post.");
+    }
+
+    await pool.query(
         'UPDATE posts SET title = ?, content = ? WHERE id = ?',
         [title, content, id]
     );
-    if (result.affectedRows === 0) {
-        throw new ApiError(404, "Post not found");
-    }
-    return getPostById(id);
+    const updatedPost = await getPostById(id);
+    return updatedPost;
 };
 
-export const deletePost = async (id) => {
-    const [result] = await pool.query('DELETE FROM posts WHERE id = ?', [id]);
-    if (result.affectedRows === 0) {
-        throw new ApiError(404, "Post not found");
+export const deletePost = async (id, userId) => {
+
+    const post = await getPostById(id);
+
+    if (post.authorId !== userId) {
+        throw new ApiError(403, "Forbidden: You do not have permission to delete this post.");
     }
-    return true;
+    
+    const [result] = await pool.query('DELETE FROM posts WHERE id = ?', [id]);
+    return result.affectedRows;
 };
 
 export const partiallyUpdatePost = async (id, updates) => {
